+++ frontend/src/styles/Formulario.css
@@ -1,0 +1,6 @@
+.formulario-page-container{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
