+++ frontend/src/styles/Inscripciones.css
@@ -84,6 +84,7 @@
     font-size: 1rem;
     border-radius: 40px;
     margin-right: 0.5%;
+    margin-bottom: 1%;
 }
 .boton-añadir-formularios:hover{
     background-color: #0077cc;
