+++ frontend/src/styles/RegistroForm.css
@@ -9,14 +9,14 @@
 }
 .input-registro{
     background-color: #ffff !important;
-    border: 0.8px solid #3498db; /* Cambia el color como quieras */
+    border: 0.8px solid #3498db !important; /* Cambia el color como quieras */
     border-radius: 10px;
     font-size: 1em;
     padding: 0.6em 0.6em 0.6em 1em;
 }
 .input-registro:focus {
     outline: none;
-    border: 2px solid #003366; /* Color del borde al enfocar */
+    border: 2px solid #003366 !important; /* Color del borde al enfocar */
     box-shadow: 0 0 3px #003366; /* Sombra opcional para dar efecto suave */
 }  
 .nombre-registro{
